Add accessible labels to the mailing list forms.
The `placeholder` attribute is [not to be used as a label](). The structure of the page makes it somewhat clear what the forms are for – visually. However, a user with lowered or no vision using a screen reader and/or zooming software will have difficulty making out which form is which. While many screen readers do read the contents of `placeholder` attributes when a label is missing, this is non-standard behavior, and the text is no longer there to support the user once they have started typing a value. Further, the text ”Enter email” may be insufficient to explain which field is focused.

I’ve added a visually hidden `label` element, with a text combining the basic action required (”Enter email”) with a distinguishing text per field (”...for the Django Users mailing list” etc), so that a user navigating between form fields will have a clear idea of which field is currently focused.